+++ src/DrawioAppServer/ConfiguredDrawioAppServer.ts
@@ -1,7 +1,7 @@
 import { DrawioAppServer } from "./DrawioAppServer";
 import { SelfHostedDrawioAppServer } from "./SelfHostedDrawioAppServer";
 import { Disposable } from "@hediet/std/disposable";
-import { Webview } from "vscode";
+import { Webview, OutputChannel } from "vscode";
 import { DrawioInstance } from "../DrawioInstance";
 import { Config } from "../Config";
 import { ExternallyHostedDrawioAppServer } from "./ExternallyHostedDrawioAppServer";
@@ -11,19 +11,26 @@ export class ConfiguredDrawioAppServer implements DrawioAppServer {
 
 	private selfHostedAppServer: SelfHostedDrawioAppServer | undefined;
 
-	constructor(public readonly config: Config) {}
+	constructor(
+		private readonly config: Config,
+		private readonly log: OutputChannel
+	) {}
 
 	setupWebview(webview: Webview): Promise<DrawioInstance> {
 		let target: DrawioAppServer;
 		if (this.config.useOfflineMode) {
 			if (!this.selfHostedAppServer) {
 				this.selfHostedAppServer = this.dispose.track(
-					new SelfHostedDrawioAppServer()
+					new SelfHostedDrawioAppServer(this.log, this.config)
 				);
 			}
 			target = this.selfHostedAppServer;
 		} else {
-			target = new ExternallyHostedDrawioAppServer(this.config.drawioUrl);
+			target = new ExternallyHostedDrawioAppServer(
+				this.config.drawioUrl,
+				this.log,
+				this.config
+			);
 		}
 
 		return target.setupWebview(webview);