+++ src/DrawioAppServer/HostedDrawioAppServer.ts
@@ -2,25 +2,19 @@ import * as vscode from "vscode";
 import { Webview } from "vscode";
 import { DrawioInstance } from "../DrawioInstance";
 import { DrawioAppServer } from "./DrawioAppServer";
+import { formatValue } from "./formatValue";
+import { Config } from "../Config";
 
 export abstract class HostedDrawioAppServer implements DrawioAppServer {
 	public abstract getIndexUrl(): Promise<string>;
 
+	constructor(
+		private readonly log: vscode.OutputChannel,
+		private readonly config: Config
+	) {}
+
 	public async setupWebview(webview: Webview): Promise<DrawioInstance> {
 		webview.options = { enableScripts: true };
-
-		let ui = "dark";
-
-		try {
-			const ctk = (vscode as any).ColorThemeKind;
-			ui = {
-				[ctk.Light]: "Kennedy",
-				[ctk.Dark]: "dark",
-				[ctk.HighContrast]: "Kennedy",
-			}[(vscode as any).window.activeColorTheme.kind];
-		} catch (e) {
-			// window.activeColorTheme is only supported since VS Code 45.
-		}
 
 		const indexUrl = await this.getIndexUrl();
 
@@ -50,16 +44,51 @@ export abstract class HostedDrawioAppServer implements DrawioAppServer {
 					});
 				</script>
 	
-				<iframe src="${indexUrl}?embed=1&ui=${ui}&proto=json&configure=1"></iframe>
+				<iframe src="${indexUrl}?embed=1&ui=${this.getTheme()}&proto=json&configure=1"></iframe>
 			</body>
 		</html>
 			`;
 
 		const drawioInstance = new DrawioInstance({
-			sendMessage: (msg) => webview.postMessage(msg),
-			registerMessageHandler: (handler) =>
-				webview.onDidReceiveMessage(handler),
+			sendMessage: (msg) => {
+				this.log.appendLine("vscode -> drawio: " + prettify(msg));
+				webview.postMessage(msg);
+			},
+			registerMessageHandler: (handler) => {
+				return webview.onDidReceiveMessage((msg) => {
+					this.log.appendLine("vscode <- drawio: " + prettify(msg));
+					handler(msg);
+				});
+			},
 		});
 		return drawioInstance;
 	}
+
+	private getTheme(): string {
+		if (this.config.drawioTheme !== "automatic") {
+			return this.config.drawioTheme;
+		}
+
+		try {
+			const ctk = (vscode as any).ColorThemeKind;
+			return {
+				[ctk.Light]: "Kennedy",
+				[ctk.Dark]: "dark",
+				[ctk.HighContrast]: "Kennedy",
+			}[(vscode as any).window.activeColorTheme.kind];
+		} catch (e) {
+			// window.activeColorTheme is only supported since VS Code 1.45.
+		}
+		return "dark";
+	}
+}
+
+function prettify(msg: unknown): string {
+	try {
+		if (typeof msg === "string") {
+			const obj = JSON.parse(msg as string);
+			return formatValue(obj, 80);
+		}
+	} catch {}
+	return "" + msg;
 }