+++ src/DrawioTextEditorProvider.ts
@@ -9,6 +9,7 @@ import {
 } from "vscode";
 import * as formatter from "xml-formatter";
 import { DrawioAppServer } from "./DrawioAppServer";
+import { canonicalizeXml } from "./canonicalizeXml";
 
 export class DrawioTextEditorProvider implements CustomTextEditorProvider {
 	constructor(public readonly drawioAppServer: DrawioAppServer) {}
@@ -21,6 +22,8 @@ export class DrawioTextEditorProvider implements CustomTextEditorProvider {
 		const drawioInstance = await this.drawioAppServer.setupWebview(
 			webviewPanel.webview
 		);
+
+		let lastOutput: string;
 		let isThisEditorSaving = false;
 
 		workspace.onDidChangeTextDocument((evt) => {
@@ -31,23 +34,66 @@ export class DrawioTextEditorProvider implements CustomTextEditorProvider {
 				// We don't want to integrate our own changes
 				return;
 			}
-			drawioInstance.loadXml(evt.document.getText());
+			if (evt.contentChanges.length === 0) {
+				// Sometimes VS Code reports a document change without a change.
+				return;
+			}
+			const result = evt.document.getText();
+			if (canonicalizeXml(result) === lastOutput) {
+				return;
+			}
+
+			drawioInstance.loadXmlLike(result);
 		});
 
 		drawioInstance.onChange.sub(async ({ newXml }) => {
 			// We format the xml so that it can be easily edited in a second text editor.
-			const formatted = formatter(newXml);
 
-			const edit = new WorkspaceEdit();
-			edit.replace(
+			let output: string;
+
+			if (document.fileName.endsWith(".svg")) {
+				const svg = await drawioInstance.exportAsSvgWithEmbeddedXml();
+				newXml = svg.toString("utf-8");
+				output = formatter(newXml);
+			} else {
+				output = formatter(newXml);
+			}
+
+			// TODO improve this.
+			lastOutput = canonicalizeXml(output);
+
+			const workspaceEdit = new WorkspaceEdit();
+
+			// TODO diff the new document with the old document and only edit the changes.
+			workspaceEdit.replace(
 				document.uri,
 				new Range(0, 0, document.lineCount, 0),
-				formatted
+				output
 			);
 
 			isThisEditorSaving = true;
-			await workspace.applyEdit(edit);
-			isThisEditorSaving = false;
+			try {
+				await workspace.applyEdit(workspaceEdit);
+
+				/*
+				// This does not work until we can use the same FormattingOptions
+				// that VS Code is using
+				const formatEdits = await commands.executeCommand<TextEdit[]>(
+					"vscode.executeFormatDocumentProvider",
+					document.uri,
+					{
+						insertSpaces: true,
+						tabSize: 4,
+					} as FormattingOptions
+				);
+				if (formatEdits !== undefined) {
+					const formatWorkspaceEdit = new WorkspaceEdit();
+					formatWorkspaceEdit.set(document.uri, formatEdits);
+					await workspace.applyEdit(formatWorkspaceEdit);
+				}*/
+			} finally {
+				isThisEditorSaving = false;
+			}
 		});
 
 		drawioInstance.onSave.sub(async () => {
@@ -55,7 +101,7 @@ export class DrawioTextEditorProvider implements CustomTextEditorProvider {
 		});
 
 		drawioInstance.onInit.one(async () => {
-			drawioInstance.loadXml(document.getText());
+			drawioInstance.loadXmlLike(document.getText());
 		});
 	}
 }