+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.3.0
+
+    - Supports editing `*.drawio.svg` files.
+    - Introduces `hediet.vscode-drawio.theme` to configure the theme used in the Draw.io editor.
+    - Logs the drawio iframe/extension communication.
+    - Fixes a memory leak.
+    - Fixes a bug that resets the view/undo stack on save.
+
 ## 0.2.0
 
 -   Implements offline mode (enabled by default).
@@ -8,7 +16,7 @@
 
 ## 0.1.3
 
--   Treats \*.dio files the same as \*.drawio files.
+-   Treats `*.dio` files the same as `*.drawio` files.
 -   Makes extension compatible with VS Code 1.44.
 
 ## 0.1.0